Do not modify images after loading them
In BSP land, every image has a header for passing parameters around, and
the TOC1 loader likes to poke at them. Stop doing that, as this
overwrites real data in mainline. Make the loader as boring as possible:
it now copies each image to whatever address in one chunk, and does
nothing else to it.

The only other bit of work the loader still does is to save off the
addresses of the more important images.

Signed-off-by: Samuel Holland <samuel@sholland.org>