Always use the load address from the TOC1 item header
The TOC1 loader expects a header inside some item images (the "image
header"), which duplicates data from the TOC1 item header. Since the
item images must be aligned to the start of a sector, this throws off
alignment of the actual image contents.

Resolve this craziness by just using the load address from the TOC1 item
header. This is backward compatible with TOC1 images created by the BSP,
because the address in the image header always matches the item header!

Also stop overriding the load address for several other image types.

Signed-off-by: Samuel Holland <samuel@sholland.org>