Add reverse-engineered C version of DRAM init code
The reverse-engineering work was done by Paul Ruizendaal, and a history
of that effort is available at https://gitlab.com/pnru/boot0.

mctl_hal.c is a direct 1:1 translation of the code in mctl_hal.S. Except
for two unused functions, which are noted in the C file, the entire
contents of mctl_hal.S were translated.

This commit imports Paul's code as-is. The only changes to the files are
whitespace and SPDX header fixes. Further commits will integrate this
code with the existing build system.

Signed-off-by: Paul Ruizendaal <pnr@planet.nl>
Signed-off-by: Samuel Holland <samuel@sholland.org>